Shm16 kozlofski - draft

+++ shm/player.cpp
@@ -1,7 +1,7 @@
 #include "player.hpp"
 
 Player::Player(const Ship& ship,
-               int money,
+               size_t money,
                size_t availableSpace)
     : ship_(std::make_shared<Ship>(ship)),
       money_(money),
@@ -12,6 +12,9 @@ size_t Player::calcAvailableSpace() {
     for (const auto el : ship_->getCargos()) {
         spaceOccupied += el->getAmount();
     }
-    availableSpace_ = ship_->getCapacity() - spaceOccupied;
+    availableSpace_ = ship_->getCapacity() - spaceOccupied;  // maybe assigning to this private field is not needed -
+    // - only calculate & return?
+    // and method would remain const, as earlier
+
     return availableSpace_;
 }
